story-generator/requirements.txt
langchain
langchain-openai
python-dotenv
fastapi 
uvicorn